## Description
TurnUpdate is not performing as expected after merge of branch. Notation and

## Expected Behavior

There should be a message with each turn update and should advance with no exceptions

## Current Behavior

An empty string message as well as a WrongPlayerException

## Steps to Reproduce

Provide detailed steps for reproducing the issue.

 1. Run an AI game with you as the white player
 2. Make the first move normally
 3. Ai responds normally, and perform second move
 4. Error thrown + history not active

## Failure Logs

Error message shown on UI saying "WrongPlayerException: Expected: 1 Actual: 0" and history not showing up

## Hypothesis for Fixing the Bug

The issue stemmed from a merge. This test checks for the message, which would indicate the faulty code from the merge has been reverted.