## Description
When the host game was being initialized, the ChessBoard object

## Expected Behavior

The host game is expected to go to the default chessboard screen on the frontend and display the starting configurations of the game

## Current Behavior

Once the game is uploaded, an exception is thrown in the command line, as defined below in the LOG section of the bug report

## Steps to Reproduce

Provide detailed steps for reproducing the issue.

 1. Upload the JSON for the accompanying game in the host game page
 2. Enter Room ID and click host game
 3. Check the logs for the error message

## Failure Logs

14:46:47.845 [JavaFX Application Thread] ERROR oogasalad.GamePlayer.Board.TurnManagement.RemoteTurnManager - Could not resolve type id 'java.util.LinkedHashMap' as a subtype of `java.util.Collection<oogasalad.GamePlayer.Board.EndConditions.EndCondition>`: Not a subtype
 at [Source: (String)"["java.util.LinkedHashMap",{"timestamp":["java.util.Date",1650998807842],"status":404,"error":"Not Found","message":"No message available","path":"/turns/getEndConditions/"}]"; line: 1, column: 28]

## Hypothesis for Fixing the Bug

The path for the APi should be "/turns/getEndConditions/{id}", however the id was not added at the end of the URL, meaning that the ID is an empty string and the initialization of the objects was unsuccessful, leading to an exception being thrown. my hypothesis is that changing the "toServerBoard" method in the ChessBoard will fix this issues.

The test added checks whether the host game was able to parse and create the game without throwing an exception.

* NOTE: the existing JacksonTests has been renamed to ServerManagerTests to better reflect the addition of the new test.