## Description
Turns were not appropriately being displayed with join game, once the game is rendered

## Expected Behavior

I should not be able to move the piece of the opposing team

## Current Behavior

I am able to move the opposing teams piece

## Steps to Reproduce

1. Create Host Game
1. Make a move for your side
2. Close and create new game instance
3. Join Remote Game
4. Make a move for your side
5. Make a move for the opposite team

## Failure Logs

No log displayed as the moves are valid for a local game.

## Hypothesis for Fixing the Bug

There is an issue with checking whether or not a move could be made during initalization of a remote game.