## Description
Thread issue with Java FX and

## Expected Behavior

When the timer goes off, an API call is made from the server.

## Current Behavior

A thread issue is thrown whenever the timer is run

## Steps to Reproduce

Provide detailed steps for reproducing the issue.

 1. Create Host game on one instance
 2. Create Join game on another instance
 3. Make move on host
 4. Make move on join
 5. Wait atleast 10 seconds and an error is thrown in the console with the move not being updates on the host game

## Failure Logs

Exception in thread "Timer-1" java.lang.IllegalStateException: Not on FX application thread; currentThread = Timer-1

## Hypothesis for Fixing the Bug

It's a thread issue. I was previosuly looking at ways to solve it without using a JavaFx class, but realized I now had no choice and will implement that.

This test just checkes if the exception is thrown